+++ src/main/java/team/awesome/pet/dao/PetRescueOrganizationDao.java
@@ -1,6 +1,5 @@
 package team.awesome.pet.dao;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import team.awesome.pet.model.Pet;
@@ -57,7 +56,7 @@ public interface PetRescueOrganizationDao {
 
   void updateShelter(Shelter shelter);
 
-  void removeShelterById(BigInteger id);
+  void removeShelterById(int id);
 
   /**
    * All User's related Restful API calls.
