+++ src/main/java/team/awesome/pet/dao/ShelterDao.java
@@ -1,6 +1,5 @@
 package team.awesome.pet.dao;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import team.awesome.pet.model.Shelter;
@@ -14,5 +13,5 @@ public interface ShelterDao {
 
   void insertShelterToDb(Shelter shelter);
 
-  void removeShelterById(BigInteger id);
+  void removeShelterById(int id);
 }
