+++ src/main/java/team/awesome/pet/service/ShelterService.java
@@ -1,6 +1,5 @@
 package team.awesome.pet.service;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,7 +31,7 @@ public class ShelterService {
     this.shelterDao.updateShelter(shelter);
   }
 
-  public void removeShelterById(BigInteger id) {
+  public void removeShelterById(int id) {
     this.shelterDao.removeShelterById(id);
   }
 }
